ShaderImageLoadStore: relax texture incompleteness requirement
When the test was writen, OpenGL spec specified that image access
would be considered invalid fs the texture is incomplete.

This raised some issues when the texture is mipmap-incomplete (so
incomplete as a whole), but base-complete, and the image is bound to
the base level.

After some debate it was agreed to allow that case.

This patch modifies the test by setting a level different to the base
level for the incomplete source/destination images.

Affects:
KHR-GL45.XXX

Components: OpenGL
Khronos bugzilla: 16111